# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio-Fixer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/images/upload:
    post:
      operationId: uploadImage
      tags: [images]
      summary: Upload an image (admin only)
      security: [{ adminAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                entityType: { type: string }
                entityId: { type: string }
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { $ref: "#/components/schemas/ImageUploadResult" }
                  - $ref: "#/components/schemas/ApiError"
  /v1/images/{id}/metadata:
    get:
      operationId: getImageMetadata
      tags: [images]
      summary: Get image metadata
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Image metadata
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { $ref: "#/components/schemas/ImageMetadata" }
                  - $ref: "#/components/schemas/ApiError"
  /v1/images/{id}:
    delete:
      operationId: deleteImage
      tags: [images]
      summary: Delete an image (admin only)
      security: [{ adminAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { type: object, nullable: true }
                  - $ref: "#/components/schemas/ApiError"
  /v1/images/reorder:
    post:
      operationId: reorderImages
      tags: [images]
      summary: Reorder images (admin only)
      security: [{ adminAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ordered_ids]
              properties:
                ordered_ids: { type: array, items: { type: string } }
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { type: object, nullable: true }
                  - $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
    ApiError:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, enum: [false] }
        message: { type: string }
        code: { type: string }
        existingId: { type: string }
...
    ImageUploadResult:
      type: object
      properties:
        id: { type: string }
        url: { type: string }
        variants:
          type: array
          items:
            type: object
            properties:
              type: { type: string }
              url: { type: string }
    ImageMetadata:
      type: object
      properties:
        id: { type: string }
        original_filename: { type: string }
        mime_type: { type: string }
        file_size_bytes: { type: integer }
        entity_type: { type: string }
        entity_id: { type: string, nullable: true }
        created_at: { type: string }
...
  securitySchemes:
    adminAuth:
      type: apiKey
      in: header
      name: x-admin-key
      description: Admin API key for server-side operations. Alternatively, use Bearer token from Clerk JWT.